% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipdown.R
\name{flipdown}
\alias{flipdown}
\title{Implement a countdown}
\usage{
flipdown(
  downto,
  id,
  theme = "dark",
  headings = c("Days", "Hours", "Minutes", "Seconds")
)
}
\arguments{
\item{downto}{the targeted date in 'yyyy-mm-dd' or 'yyyy-mm-dd hh:mm:ss' format}

\item{id}{the id of the flipdown object}

\item{theme}{the theme of the flipdown object. Two possible choices: 'dark' or 'light'. Defaults to 'dark'}

\item{headings}{a character vector specifying the heading related to each time part. Defaults to English headings: c("Days", "Hours", "Minutes", "Seconds")}
}
\value{
a countdown object
}
\description{
Implement a countdown
}
\examples{

flipdown(downto = "2020-11-16 20:20:20 UTC", id = "first", theme = "dark")


}
